Update Vagrantfile
New MS links to Vagrant are zip files containing the .box (instead of downloading the .box directly). Because of this change, Vagrant would not correct import the content of the .box.

We changed the behavior of this Vagrantfile so that it would look for the proper modern.ie box, if not imported, will download, extract and import (using regular bash command).